Revert adding a OneBox stage for now
If the OneBox and Prod are in the same target groups, the CodeDeploy OneBox deployment will move OneBox to 'green' and flip prod traffic to it.  But Prod tasks are still in 'blue', resulting in *only* OneBox serving prod traffic.

Better solution is to use different target groups + weighted traffic routing, but that feature hasn't shown up in CDK yet